+++ web/templates/about.html
@@ -6,9 +6,17 @@
 {% block body %}
 <h1 class="page-header">About</h1>
 <p class="lead">
-Raspberry Pi Web Camera runs on Rasberry PI 2 <a href="http://www.raspberrypi.org/">Rasberry PI 2</a>
-and uses <a href="http://www.raspberrypi.org/help/camera-module-setup/">Raspberry PI camera</a>.
+&emsp; Smart Home Solutions was created to fill in a void created by the market.
+The project chose to focus on the aspect of facial recognition and geofencing technologies.
+and bring them into the smart home available today.
 </p>
+<ul class="list-group">
+Team Members:
+<p>Sukhjot Bassi</p>
+<p>Jeremy Chau</p>
+<p>Sophia Quan</p>
+<p>Rizen Yamauchi</p>
+</ul>
 <div class="row">
   <div class="col-md-4">
     <ul class="list-group">
