+++ web/templates/home.html
@@ -6,17 +6,18 @@
 {% block body %}
 <h1 class="page-header">Home</h1>
 <div class="jumbotron">
-  <h1>Raspberry Pi Web Camera</h1>
-  <p align="center"></p>
   <div class="row">
     <div class="col-md-3"></div>
-    <div class="col-md-6"><p align="center">Live stream, motion detection, recordings.<p></div>
+    <div class="col-md-6"><p align="center">Smart Home Solutions<p></div>
+    <div class="col-md-3"></div>
+    <div class="col-md-3"></div>
+    <div class="col-md-6"><p align="center"><p></div>
     <div class="col-md-3"></div>
   </div>
   <div class="row">
     <div class="col-md-4"></div>
-    <div class="col-md-4"><p align="center"><a class="btn btn-primary btn-lg" href="{% url 'home.views.about' %}" role="button">Learn more</a></p></div>
-    <div class="col-md-4"><p align="right"><img src="{% static 'img/logo_256.png' %}"></p></div>
+    <div class="col-md-4"><p align="center"><a class="btn btn-info btn-lg" href="{% url 'home.views.about' %}" role="button">Learn more</a></p></div>
+    <div class="col-md-4"><p align="center"><img src="{% static 'img/Smart_Home_Logo.png' %}"></p></div>
   </div>
 </div>
 {% endblock %}
